requirements/local.txt
# Local development dependencies go here
-r production.txt
mock==2.0.0
coverage==3.7.1
django-discover-runner==1.0
django-extensions==1.3.3
Sphinx

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.0.1
